shuffleboard is kill
